processing-docker/requirements.txt
celery[redis]==4.4.6
pandas==1.0.3
requests-toolbelt==0.9.1
numpy==1.17.4
scikit-stats==0.1.2
matplotlib==3.3.4
python-decouple==3.6
